change data format for TOT to accomodate value over 255

+++ include/EUTelSimpleSparsePixel.h
@@ -96,7 +96,7 @@ namespace eutelescope {
     short _yCoord;
 
     //! The signal
-    short _signal;
+    int _signal;
 
   };
 }
